Full STARTROPICS sync. Get a bunch of stuff compiling again (recent and older cc-lib changes); fix some compiler warnings; maybe some stray files I forgot to check in.
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@5217 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ fceulib/aot.cc
@@ -128,7 +128,7 @@ struct Code {
   vector<uint8> code;
 
   uint8 Get(int addr) const {
-    CHECK(addr >= 0 && addr < code.size() && known[addr]) << "Unmapped "
+    CHECK(addr >= 0 && addr < (int)code.size() && known[addr]) << "Unmapped "
       "read from code: " << addr;
     return code[addr];
   }
